requirements.txt
discord.py==2.2.2
python-dotenv==0.15.0
